+++ test/runtime/frontend_test/chainer_test/sub_from_constant_test.py
@@ -6,7 +6,7 @@ from webdnn.frontend.chainer.converter import ChainerConverter
 
 
 def test():
-    vx = chainer.Variable(np.random.rand(2, 4, 6, 8))
+    vx = chainer.Variable(np.random.rand(2, 4, 6, 8).astype(np.float32))
     vy = vx - 1
 
     graph = ChainerConverter().convert([vx], [vy])
@@ -23,7 +23,7 @@ def test():
 
 
 def test_rsub():
-    vx = chainer.Variable(np.random.rand(2, 4, 6, 8))
+    vx = chainer.Variable(np.random.rand(2, 4, 6, 8).astype(np.float32))
     vy = 1 - vx
 
     graph = ChainerConverter().convert([vx], [vy])
