+++ test/runtime/frontend_test/tensorflow_test/ops_test/gen_math_ops_test/add_n_test.py
@@ -10,8 +10,8 @@ def template(x0_shape, x1_shape, description: str = ""):
     x1 = tf.placeholder(np.float32, x1_shape, "x1")
     y = tf.add_n([x0, x1])
 
-    vx0 = np.random.rand(*x0_shape)
-    vx1 = np.random.rand(*x1_shape)
+    vx0 = np.random.rand(*x0_shape).astype(np.float32)
+    vx1 = np.random.rand(*x1_shape).astype(np.float32)
     with tf.Session() as sess:
         vy, = sess.run([y], {x0: vx0, x1: vx1})
